+++ mst/mplatform.h
@@ -35,12 +35,12 @@ namespace platform {
 
 inline ::std::string name() noexcept
 {
-	return ::mst::platform::_Details::get_os_name_impl();
+	return ::mst::_Details::get_os_name_impl();
 }
 
 inline ::std::string version() noexcept
 {
-	return ::mst::platform::_Details::get_os_version_string_impl();
+	return ::mst::_Details::get_os_version_string_impl();
 }
 
 inline ::std::string full_name() noexcept
@@ -54,6 +54,8 @@ inline ::std::string newline() noexcept
 	return "\r\n";
 #elif MST_PLATFORM_LINUX
 	return "\n";
+#elif MST_PLATFORM_MAC
+	return "\r";
 #else
 #error "Platform not implemented"
 #endif
@@ -65,6 +67,8 @@ inline char directory_separator() noexcept
 	return '\\';
 #elif MST_PLATFORM_LINUX
 	return '/';
+#elif MST_PLATFORM_MAC
+	return '/';
 #else
 #error "Platform not implemented"
 #endif
@@ -73,7 +77,7 @@ inline char directory_separator() noexcept
 inline ::std::string downloads_path() noexcept
 {
 	char path[1024];
-	if(mst::platform::_Details::get_downloads_folder_impl(path))
+	if(::mst::_Details::get_downloads_folder_impl(path))
 	{
 		return path;
 	}
@@ -83,7 +87,7 @@ inline ::std::string downloads_path() noexcept
 inline ::std::string desktop_path() noexcept
 {
 	char path[1024];
-	if(mst::platform::_Details::get_desktop_folder_impl(path))
+	if(::mst::_Details::get_desktop_folder_impl(path))
 	{
 		return path;
 	}
@@ -93,7 +97,7 @@ inline ::std::string desktop_path() noexcept
 inline ::std::string my_documents_path() noexcept
 {
 	char path[1024];
-	if(mst::platform::_Details::get_mydocuments_folder_impl(path))
+	if(::mst::_Details::get_mydocuments_folder_impl(path))
 	{
 		return path;
 	}
@@ -103,7 +107,7 @@ inline ::std::string my_documents_path() noexcept
 inline ::std::string temp_path() noexcept
 {
 	char path[1024];
-	if(mst::platform::_Details::get_temp_folder_impl(path))
+	if(::mst::_Details::get_temp_folder_impl(path))
 	{
 		return path;
 	}
@@ -113,7 +117,7 @@ inline ::std::string temp_path() noexcept
 inline ::std::string recycle_bin_path() noexcept
 {
 	char path[1024];
-	if(mst::platform::_Details::get_recycle_bin_folder_impl(path))
+	if(::mst::_Details::get_recycle_bin_folder_impl(path))
 	{
 		return path;
 	}
@@ -122,13 +126,13 @@ inline ::std::string recycle_bin_path() noexcept
 
 inline bool create_directory(const ::std::string& path) noexcept
 {
-	return mst::platform::_Details::create_directory_impl(path.c_str());
+	return ::mst::_Details::create_directory_impl(path.c_str());
 }
 
 inline ::std::string current_directory() noexcept
 {
 	char path[1024];
-	if(mst::platform::_Details::get_current_directory_impl(path))
+	if(::mst::_Details::get_current_directory_impl(path))
 	{
 		return path;
 	}
@@ -138,22 +142,22 @@ inline ::std::string current_directory() noexcept
 inline bool set_current_directory(const ::std::string& path) noexcept
 {
 	// true if success, false if failed
-	return mst::platform::_Details::set_current_directory_impl(path.c_str());
+	return ::mst::_Details::set_current_directory_impl(path.c_str());
 }
 
 inline uint32_t page_size() noexcept
 {
-	return ::mst::platform::_Details::get_page_size_impl();
+	return ::mst::_Details::get_page_size_impl();
 }
 
 inline uint32_t processor_core_count() noexcept
 {
-	return ::mst::platform::_Details::get_processor_core_count_impl();
+	return ::mst::_Details::get_processor_core_count_impl();
 }
 
 inline uint32_t processor_thread_count() noexcept
 {
-	return ::mst::platform::_Details::get_processor_thread_count_impl();
+	return ::mst::_Details::get_processor_thread_count_impl();
 }
 
 enum class processor_feature_flags : uint32_t
@@ -209,7 +213,7 @@ enum class processor_feature_flags : uint32_t
 
 inline mst::flag<processor_feature_flags> processor_features() noexcept
 {
-	return _Details::processor_features_impl();
+	return mst::flag<processor_feature_flags>::from_storage(_Details::processor_features_impl());
 }
