+++ tests/containers/sparse_set.cpp
@@ -23,6 +23,7 @@
 //																							//
 //////////////////////////////////////////////////////////////////////////////////////////////
 
+#include <limits>
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
